starting framework for decent concurrency
