+++ crates/ploke-tui/Cargo.toml
@@ -11,6 +11,7 @@ crossterm = { version = "0.29", features = ["event-stream", "serde"] }
 tokio = { workspace = true }                                           # "full" for simplicity, you can narrow down later
 tokio-utils = "0.1"
 flume = { workspace = true }
+num_cpus = "1.0"
 # network
 reqwest = { version = "0.11", features = ["json"] }
 # common
