+++ crates/ploke-tui/src/chat_history.rs
@@ -1,7 +1,7 @@
 use super::*;
 use std::fmt;
 
-#[derive(Debug)]
+#[derive(Debug, Clone, Copy)]
 pub enum NavigationDirection {
     Next,
     Previous,
